feat(ui): 顶栏瘦身为两行小字
- 顶栏改为「品牌名 · 客服 Agent」+ 一行 muted 状态摘要
  (运行中 · 通道连接 · 刷新时间),右侧只留主题切换;
  去掉状态灯带与大字号品牌。
- 页头主操作按钮恢复默认尺寸,与参考实现的页头按钮一致。
- 随之删除只服务旧顶栏的 brand-title / channel-status-lights /
  polling-indicator;配置页要用的单通道状态点拆到 channel-dot。
